examples/wagtaildemo/requirements/production.txt
-r base.txt
django-celery==3.1.1
gunicorn==18.0
redis==2.9.1
django-redis-cache==0.10.2
elasticsearch==1.0.0
